set up voting
